requirements_dev.txt
aiohttp==3.8.3 ; python_version >= "3.9" and python_version < "3.11"
aiosignal==1.2.0 ; python_version >= "3.9" and python_version < "3.11"
async-timeout==4.0.2 ; python_version >= "3.9" and python_version < "3.11"
attrs==22.1.0 ; python_version >= "3.9" and python_version < "3.11"
charset-normalizer==2.1.1 ; python_version >= "3.9" and python_version < "3.11"
frozenlist==1.3.1 ; python_version >= "3.9" and python_version < "3.11"
graphql-subscription-manager==0.6.1 ; python_version >= "3.9" and python_version < "3.11"
idna==3.4 ; python_version >= "3.9" and python_version < "3.11"
multidict==6.0.2 ; python_version >= "3.9" and python_version < "3.11"
python-dateutil==2.8.2 ; python_version >= "3.9" and python_version < "3.11"
pytibber==0.25.3 ; python_version >= "3.9" and python_version < "3.11"
six==1.16.0 ; python_version >= "3.9" and python_version < "3.11"
websockets==10.3 ; python_version >= "3.9" and python_version < "3.11"
yarl==1.8.1 ; python_version >= "3.9" and python_version < "3.11"
